Refactor TV action to use new home screen image
The TV action in the room navigation screen has been updated to use the
new home screen image for the TV interface. This change aligns the image
correctly and sets the appropriate size. The previous image had
inconsistent alignment and size, which affected the user experience.
This update improves the visual presentation of the TV interface. Fixes
#1234.